automation/requirements.txt
playwright==1.52.0
psycopg[binary]==3.2.6
PyYAML==6.0.2
